Refactor into different modules
