+++ app/views/modules/list.scala.html
@@ -1,6 +1,4 @@
-@import models.modules._
-
-@(modules: Seq[Module])(implicit requestHeader : RequestHeader)
+@(modules: Seq[models.modules.Module])(implicit requestHeader : RequestHeader)
 
 @layout {
