+++ modules/certification/api/src/main/scala/models/certification/Certification.scala
@@ -1,0 +1,8 @@
+package models.certification
+
+import org.joda.time.DateTime
+
+/**
+ * Interest in a certification
+ */
+case class Certification(registered: DateTime, name: String, email: String, developer: Boolean, organization: Boolean, comments: String)
