+++ modules/certification/api/src/main/scala/services/certification/CertificationDao.scala
@@ -1,0 +1,16 @@
+package services.certification
+
+import models.certification.Certification
+
+trait CertificationDao {
+
+  /**
+   * Someone has registered interest in certification
+   */
+  def registerInterest(certification: Certification): Unit
+
+  /**
+   * Find all people that have registered interest in certification
+   */
+  def findAllInterested(): Seq[Certification]
+}
