+++ modules/certification/db/build.sbt
@@ -1,0 +1,4 @@
+
+Common.settings
+
+libraryDependencies ++= Common.libraryDependencies
