+++ modules/module/db/build.sbt
@@ -1,0 +1,5 @@
+
+
+Common.settings
+
+libraryDependencies ++= Common.libraryDependencies
