+++ modules/utils/build.sbt
@@ -1,0 +1,11 @@
+
+
+Common.settings
+
+libraryDependencies ++= Common.libraryDependencies
+
+libraryDependencies ++= Seq(
+  "com.typesafe.play" %% "play-doc" % "1.5.0",
+  "org.eclipse.jgit" % "org.eclipse.jgit" % "3.0.0.201306101825-r",
+  "org.fusesource.wikitext" % "textile-core" % "1.3"
+)
